requirements.txt
pytest==4.2.1
PyYAML==3.13